+++ src/edu/ucsb/cs56/drawings/sswong/advanced/Tree.java
@@ -1,4 +1,4 @@
-package edu.ucsb.cs56.drawings.sswong;
+package edu.ucsb.cs56.drawings.sswong.advanced;
 import java.awt.geom.GeneralPath; // combinations of lines and curves
 import java.awt.Shape; // general class for shapes
